Moved the complex tests on Throttled into "MessageLoopTests" and expanded on the simple ones
