+++ JustSaying.Messaging.UnitTests/JustSaying.Messaging.UnitTests.csproj
@@ -69,6 +69,7 @@
     <Reference Include="Microsoft.CSharp" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="MessageProcessingStrategies\MessageLoopTests.cs" />
     <Compile Include="MessageProcessingStrategies\ThrottledTests.cs" />
     <Compile Include="MessageHandling\WhenEnsuringMessageIsOnlyHandledExactlyOnce.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
