+++ JustSaying.Messaging.UnitTests/MessageProcessingStrategies/ThrottledTests.cs
@@ -1,8 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Linq;
-using System.Threading;
 using System.Threading.Tasks;
 using JustSaying.Messaging.MessageProcessingStrategies;
 using JustSaying.Messaging.Monitoring;
@@ -14,104 +10,116 @@ namespace JustSaying.Messaging.UnitTests.MessageProcessingStrategies
     [TestFixture]
     public class ThrottledTests
     {
-        private int _concurrencyLevel;
-        private Throttled _messageProcessingStrategy;
-        private int _actionsProcessed;
-        private int _fakeAmazonBatchSize;
         private IMessageMonitor _fakeMonitor;
 
         [SetUp]
         public void SetUp()
         {
             _fakeMonitor = Substitute.For<IMessageMonitor>();
-            _fakeAmazonBatchSize = 10;
-            _concurrencyLevel = 20;
-            _messageProcessingStrategy = new Throttled(_concurrencyLevel, _fakeMonitor);
-            _actionsProcessed = 0;
         }
 
         [Test]
         public void ChangeMaxAllowedMessagesInFlightAtRuntime_TheChangeIsApplied()
         {
-            var maxAllowedMessagesInFlight = Substitute.For<Func<int>>();
+            Func<int> maxAllowedMessagesInFlight = Substitute.For<Func<int>>();
             maxAllowedMessagesInFlight().Returns(100);
-            _messageProcessingStrategy = new Throttled(maxAllowedMessagesInFlight, _fakeMonitor);
+            var messageProcessingStrategy = new Throttled(maxAllowedMessagesInFlight, _fakeMonitor);
 
-            Assert.That(_messageProcessingStrategy.FreeTasks, Is.EqualTo(100));
+            Assert.That(messageProcessingStrategy.FreeTasks, Is.EqualTo(100));
 
             maxAllowedMessagesInFlight().Returns(90);
 
-            Assert.That(_messageProcessingStrategy.FreeTasks, Is.EqualTo(90));
+            Assert.That(messageProcessingStrategy.FreeTasks, Is.EqualTo(90));
         }
-
-        [TestCase(1000)]
-        [TestCase(10000)]
-        public async Task SimulatedListenLoop_ProcessedAllMessages(int numberOfMessagesToProcess)
-        {
-            var watch = new Stopwatch();
-            watch.Start();
-            var actions = BuildFakeIncomingMessages(numberOfMessagesToProcess);
-
-            await ListenLoopExecuted(actions);
-
-            watch.Stop();
-            Thread.Sleep(2000);
-
-            Assert.That(_actionsProcessed, Is.EqualTo(numberOfMessagesToProcess));
-            Debug.WriteLine("Took " + watch.Elapsed + " to process " + numberOfMessagesToProcess + " messages.");
-        }
-
 
         [Test]
-        public async Task SimulatedListenLoop_WhenThrottlingOccurs_CallsMessageMonitor()
+        public void FreeTaskCountStartsAtCapacity()
         {
-            var actions = BuildFakeIncomingMessages(50);
-            _messageProcessingStrategy = new Throttled(20, _fakeMonitor);
-
-            await ListenLoopExecuted(actions);
-
-            _fakeMonitor.Received().IncrementThrottlingStatistic();
+            var messageProcessingStrategy = new Throttled(123, _fakeMonitor);
+            Assert.That(messageProcessingStrategy.FreeTasks, Is.EqualTo(123));
         }
 
-        private async Task ListenLoopExecuted(Queue<Action> actions)
+        [Test]
+        public void WhenATasksIsAddedTheFreeTaskCountIsDecremented()
         {
-            while (actions.Any())
-            {
-                await _messageProcessingStrategy.AwaitAtLeastOneTaskToComplete();
-                var batch = GetFromFakeSnsQueue(actions);
+            var messageProcessingStrategy = new Throttled(123, _fakeMonitor);
+            var tcs = new TaskCompletionSource<bool>();
 
-                foreach (var action in batch)
-                {
-                    _messageProcessingStrategy.ProcessMessage(action);
-                }
-            }
+            messageProcessingStrategy.ProcessMessage(async () => await tcs.Task);
+
+            Assert.That(messageProcessingStrategy.FreeTasks, Is.EqualTo(122));
         }
 
-        private IEnumerable<Action> GetFromFakeSnsQueue(Queue<Action> actions)
+        [Test]
+        public async Task WhenATaskCompletesTheFreeTaskCountIsIncremented()
         {
-            var batch = new List<Action>();
-            for (var i = 0; i != _fakeAmazonBatchSize; i++)
-            {
-                batch.Add(actions.Dequeue());
-            }
-            return batch;
+            var messageProcessingStrategy = new Throttled(3, _fakeMonitor);
+            var tcs = new TaskCompletionSource<bool>();
+
+            messageProcessingStrategy.ProcessMessage(async () => await tcs.Task);
+
+            Assert.That(messageProcessingStrategy.FreeTasks, Is.EqualTo(2));
+
+            await AllowTasksToComplete(tcs);
+
+            Assert.That(messageProcessingStrategy.FreeTasks, Is.EqualTo(3));
         }
 
-        private Queue<Action> BuildFakeIncomingMessages(int numberOfMessagesToCreate, int maximumDurationASingleMessageWillTakeToExecute = 10)
+        [Test]
+        public void FreeTaskCountCanReachZero()
         {
-            var actions = new Queue<Action>();
-            for (var i = 0; i != numberOfMessagesToCreate; i++)
+            const int capacity = 10;
+            var messageProcessingStrategy = new Throttled(capacity, _fakeMonitor);
+            var tcs = new TaskCompletionSource<bool>();
+
+            for (int i = 0; i < capacity; i++)
             {
-                var random = new Random();
-                var action = new Action(() =>
-                {
-                    Thread.Sleep(random.Next(maximumDurationASingleMessageWillTakeToExecute));
-                    Interlocked.Increment(ref _actionsProcessed);
-                });
-                actions.Enqueue(action);
+                messageProcessingStrategy.ProcessMessage(async () => await tcs.Task);
             }
 
-            return actions;
+            Assert.That(messageProcessingStrategy.FreeTasks, Is.EqualTo(0));
+            tcs.SetResult(true);
+        }
+
+        [Test]
+        public async Task FreeTaskCountCanGoToZeroAndBackToFull()
+        {
+            const int capacity = 10;
+            var messageProcessingStrategy = new Throttled(capacity, _fakeMonitor);
+            var tcs = new TaskCompletionSource<bool>();
+
+            for (int i = 0; i < capacity; i++)
+            {
+                messageProcessingStrategy.ProcessMessage(async () => await tcs.Task);
+            }
+
+            Assert.That(messageProcessingStrategy.FreeTasks, Is.EqualTo(0));
+
+            await AllowTasksToComplete(tcs);
+
+            Assert.That(messageProcessingStrategy.FreeTasks, Is.EqualTo(capacity));
+        }
+
+        [Test]
+        public void FreeTaskCountIsNeverNegative()
+        {
+            const int capacity = 10;
+            var messageProcessingStrategy = new Throttled(capacity, _fakeMonitor);
+            var tcs = new TaskCompletionSource<bool>();
+
+
+            for (int i = 0; i < capacity; i++)
+            {
+                messageProcessingStrategy.ProcessMessage(async () => await tcs.Task);
+                Assert.That(messageProcessingStrategy.FreeTasks, Is.GreaterThanOrEqualTo(0));
+            }
+            tcs.SetResult(true);
+        }
+
+        private static async Task AllowTasksToComplete(TaskCompletionSource<bool> tcs)
+        {
+            tcs.SetResult(true);
+            await Task.Delay(250);
         }
     }
 }
